AP-5429: Move methods to concern
Move set_transaction_types to concern as it is used in 3 places and
import whenever it is used.
