+++ app/controllers/providers/means/check_income_answers_controller.rb
@@ -1,7 +1,8 @@
 module Providers
   module Means
     class CheckIncomeAnswersController < ProviderBaseController
-      before_action :set_transaction_types, only: :show
+      include TransactionTypeSettable
+
       helper_method :display_employment_income?
 
       def show
@@ -12,18 +13,6 @@ module Providers
       def update
         legal_aid_application.provider_assess_means!
         continue_or_draft
-      end
-
-    private
-
-      def set_transaction_types
-        @credit_transaction_types = if legal_aid_application.client_uploading_bank_statements?
-                                      TransactionType.credits.without_disregarded_benefits.without_benefits
-                                    else
-                                      TransactionType.credits.without_housing_benefits
-                                    end
-
-        @debit_transaction_types = TransactionType.debits
       end
     end
   end
